app/serializer.py
from rest_framework import serializers
from .models import Contact, ContactsMapping, Account
from django.contrib.auth.models import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = "__all__"


class ContactSerializer(serializers.ModelSerializer):
    class Meta:
        model = Contact
        fields = "__all__"


class ContactsMappingSerializer(serializers.ModelSerializer):
    class Meta:
        model = ContactsMapping
        fields = "__all__"


class AccountSerializer(serializers.ModelSerializer):
    class Meta:
        model = Account
        fields = "__all__"
