requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.12.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
Django==4.1.4
django-environ==0.9.0
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
djangorestframework==3.14.0
install==1.3.5
mypy-extensions==0.4.3
pathspec==0.10.3
phonenumbers==8.13.2
platformdirs==2.6.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.4.0
